+++ packages/licenses/src/index.ts
@@ -1,0 +1,9 @@
+#!/usr/bin/env node
+import commander from 'commander'
+import gatherLicensesCommand from './actions/gatherLicenses/command'
+;(() => {
+  const comms = commander
+  gatherLicensesCommand(comms)
+  // @ts-ignore
+  comms.parse(process.argv)
+})()