+++ packages/playground/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [10.2.0](https://vfuk-digital.visualstudio.com/Digital/_git/lib-web-aim/compare/v10.1.8...v10.2.0) (2023-06-20)
+
+**Note:** Version bump only for package @vfuk/lib-aim-playground
+
+## [10.1.8](https://vfuk-digital.visualstudio.com/Digital/_git/lib-web-aim/compare/v10.1.7...v10.1.8) (2023-03-20)
+
+**Note:** Version bump only for package @vfuk/lib-aim-playground
+
 ## [10.1.7](https://vfuk-digital.visualstudio.com/Digital/_git/lib-web-aim/compare/v10.1.6...v10.1.7) (2023-03-14)
 
 **Note:** Version bump only for package @vfuk/lib-aim-playground